Remove additional dependency and cenralise all mapping logic
